requirements.txt
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
mysql-connector-python==9.3.0
numpy==2.2.5
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
